requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
packaging==24.0
pyTelegramBotAPI==4.18.0
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1